rearrage show tests
